\name{returnALetter-package}
\alias{returnALetter-package}
\alias{returnALetter}
\docType{package}
\title{
\packageTitle{returnALetter}
}
\description{
\packageDescription{returnALetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{returnALetter}
\packageIndices{returnALetter}

}
\author{
\packageAuthor{returnALetter}

Maintainer: \packageMaintainer{returnALetter}
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{

}
\examples{
alph(2)
}
